iOS: update fcm token provider
