+++ iosHyperskillApp/iosHyperskillApp/Sources/Frameworks/Core/Injection/AppGraph+DefaultInstances.swift
@@ -1,22 +1,6 @@
 import Foundation
 import shared
 
-class IOSAppComponentImpl: shared.SharedIOSAppComponentImpl {
-    override func getIosFCMTokenProvider() -> IosFCMTokenProvider {
-        IosFCMTokenProviderImpl()
-    }
-}
-
-enum AppGraphBridge {
-    static let sharedAppGraph: iOSAppComponent = IOSAppComponentImpl(
-        userAgentInfo: UserAgentBuilder.userAgentInfo,
-        buildVariant: BuildVariant.current,
-        sentryManager: SentryManager.shared
-    )
-}
-
-// MARK: - Default Instances -
-
 extension AnalyticInteractor {
     static var `default`: AnalyticInteractor {
         AppGraphBridge.sharedAppGraph.analyticComponent.analyticInteractor
