+++ iosHyperskillApp/iosHyperskillApp/Sources/Frameworks/Core/Injection/IosFCMTokenProviderImpl.swift
@@ -1,0 +1,12 @@
+import FirebaseMessaging
+import Foundation
+import shared
+
+final class IosFCMTokenProviderImpl: IosFCMTokenProvider {
+    func getFCMToken() async throws -> String? {
+        guard Messaging.messaging().apnsToken != nil else {
+            return nil
+        }
+        return try await Messaging.messaging().token()
+    }
+}
