+++ shared/src/iosMain/kotlin/org/hyperskill/app/core/injection/IosAppComponent.kt
@@ -2,6 +2,6 @@ package org.hyperskill.app.core.injection
 
 import org.hyperskill.app.notification.remote.data.repository.IosFCMTokenProvider
 
-interface iOSAppComponent : AppGraph {
+interface IosAppComponent : AppGraph {
     fun getIosFCMTokenProvider(): IosFCMTokenProvider
 }
